Read JSON File Containing List of Strings to Test
Toolbox now contains a button that when clicked will prompt the user to
upload a JSON file containing a list of strings to sequentially test on
the automaton. As of right now, the tests are run automatically after parsing
the JSON file. Will inform the user if an error occurs, but the error cases
and messages will likely need to be revisited to make sure they're accurate.
Result of each test is printed to console.