Use Floating Panel to Display Tests
Now uses the floating panel function as the container for displaying the
test strings and results, the buttons to upload the JSON file containing
the tests, and the buttons to run them (not yet implemented). The UI was
also changed to more closely match the prototype in the associated issue
conversation.