Display Check or X if Test Passed or Failed
After a test runs, a green checkmark or a red X will be displayed to demonstrate
if the test got the expected result or not, respectively. For example, if
the test string was expected to be Accepted and was Accepted, then the test
passed, so a green checkmark will appear. However, if the test string was
expected to be Accepted and got Rejected instead, then the test failed so
a red X will appear. Same conditions with expecting Rejected. If expecting
Rejected and got Rejected, then green check mark. If expecting Rejected
but got Accepted, then red X.